requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
chromedriver-autoinstaller==0.2.2
click==7.1.2
Django==3.1.4
django-scheduler==0.9.3
fake-useragent==0.1.11
gTTS==2.2.1
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
icalendar==4.0.7
idna==3.1
libroosx @ file:///C:/Users/maxim/Downloads/libroosx-0.4.7-py3-none-any.whl
Pillow==8.1.0
playsound==1.2.2
pyarmor==6.6.1
pycryptodomex==3.10.1
pygame==2.0.1
pypiwin32==223
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pywin32==300
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2
SQLAlchemy==1.3.23
sqlparse==0.4.1
tricks-librus==0.9
urllib3==1.26.3
win10toast==0.9
